+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openvideo/core
 
+## 1.0.2
+
+### Patch Changes
+
+- Fixed playback and seeking
+
 ## 1.0.1
 
 ### Patch Changes