+++ packages/timeline/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @openvideo/timeline
 
+## 1.0.2
+
+### Patch Changes
+
+- Fixed playback and seeking
+- Updated dependencies
+  - @openvideo/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes